rework docs, add section for `#[pre/postfix]`
The crate documentation and the README were both getting pretty long, in
order to provide proper information about all of the available
attributes. The primary change here is to collapse the documentation
with `<details>` tags so that users can quickly view exactly what they
want to.

The README has also been changed to no longer be nearly identical to the
crate documentation; it now refers to docs.rs for more detailed
information.

This also closes #1, which was waiting on documentation to be added.